fix(sponsorkit): patch opencollective auth header
